final_project/dangi/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
boto3==1.34.131
botocore==1.34.131
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.1.4
debugpy==1.6.7
decorator==5.1.1
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
executing==1.2.0
flatbuffers==24.3.25
gast==0.5.4
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.7
ipykernel==6.25.0
ipython==8.14.0
jedi==0.19.0
jmespath==1.0.1
jupyter_client==8.3.0
jupyter_core==5.3.1
keras==2.15.0
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mysqlclient==2.2.4
ml-dtypes==0.2.0
nest-asyncio==1.5.7
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.25.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==1.7.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pywin32==306
pyzmq==25.1.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.2
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
traitlets==5.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.6
Werkzeug==3.0.3
wrapt==1.14.1
